requirements.txt
tensorflow>=2.5.0
keras>=2.5.0
opencv-contrib-python>=4.5.3
Pillow>=8.3.1
SpeechRecognition>=3.8.1
gTTS>=2.2.3
playsound>=1.3.0
numpy>=1.19.5
matplotlib>=3.4.3
pickle-mixin>=1.0.2
transformers>=4.11.0
torch>=1.9.0 